增加Layout从properties中初始化

+++ include/slog/base/fixed_buffer.h
@@ -48,7 +48,7 @@ private:
   static void CookieStart();
   static void CookieEnd();
 
-  int pos_;
+  size_t pos_;
   char data_[SIZE];
 };
